Fix code bug in AbstractVectorContainer, inherit ScrollArea from this class. AbstractSlider is now a template class. Fine tune ColorWheel and add class ColorSelector.

+++ src/Gui/Button.cpp
@@ -88,8 +88,8 @@ namespace BlendInt {
 			}
 
 			case WidgetRoundCornerRadius: {
-				const float* radius_p =
-								static_cast<const float*>(request.data());
+				const int* radius_p =
+								static_cast<const int*>(request.data());
 				UpdateTextPosition(size(), round_corner_type(), *radius_p,
 								text());
 				glBindVertexArray(m_vao);
